requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
decouple==0.0.7
dj-rest-auth==6.0.0
Django==5.1
django-allauth==64.1.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.7
mysqlclient==2.2.4
openai==1.42.0
packaging==24.1
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
